Refactor iqk: Factor out GEMM for legacy quants (AVX2/AVX512)
